First commit to this repository
Before now I've just been editing this lambda funtion
on AWS's inline web code editor, moving it to github
for version control and issue tracking.

The code is horribly organized, as you might expect
from an app that was slowly added to over several
months on a webpage inline editor.