Factored dynamodb functionality into its own file
Previously we had a lot of duplicated code for managing
the todo and agenda databases, since they were so similar.
Now all functions that need to interact with dynamodb are found
in wrt_dynamodb_hanlder.py. There's one function for writing, one
for reading, and one for deleting items

Fun fact, none of our databases actually delete anything, ever.
They just prepend the string [deleted] to the text item, and
the readers are configured to not fetch results that start with
[deleted]. Some may say that this is because I couldn't figure
out the API call for deleting items from dynamodb. They would
be right.