Added basic logging stuff
It now logs the parameters sent into event (at least, the user id, user
name, command, team domain, and text). This is meant to make it easier
to reproduce errors that are reported. The logs are viewable only on
cloudwatch, which is on a personal account, so no one else can view
logs, but that's ok for now.

Should in the future look into moving the lambda function off of the
personal account and into one controlled by the team. Very not mission
critical, but it'd suck for this feature creeping app to just die all at
once when I graduate.