Change makefile to specify files to be zipped
Previously, the makefile used the * wildcard to zip all the python files
in the directory. This wasn't optimal in that if we ever put other
python files in this directory (tests, for example), they'd be zipped
into the archive that gets uploaded to AWS.

Other advantage being that now if you clone the repo and then try to
make (without obtaining oauth_secret.py, the untracked file that
contains our oauth tokens), zip will issue a warning about the missing
file.