Give Aaron a new way to annoy Jacob
Now, if you submit a tagged todo item with a username in it (either
starting with @ as in "@jacob" or surrounded by parens as in "(jacob)",
the slack user with that Username will receive a notification from
slackbot telling them that they have a new todo item

Sorry for the non-descriptive first line there. I'm on a train, which
means you guys get bad commit messages.